requirements.txt
Flask
Flask-SocketIO
spherov2
python-dotenv
requests
openai 